Add runtime and dashboard hardening updates across backend and frontend.
Consolidate CI follow-up execution, ingress/timeline processing, and nested live-updating dashboard interactions so PR remediation and operator workflows remain deterministic and seamless.

+++ apps/backend/src/agent1/db/models.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from datetime import datetime
 from datetime import timezone
 
+from sqlalchemy import BigInteger
 from sqlalchemy import Boolean
 from sqlalchemy import DateTime
 from sqlalchemy import Enum
@@ -351,7 +352,7 @@ class CommentTargetModel(Base):
     issue_number: Mapped[int | None] = mapped_column(Integer, nullable=True)
     pr_number: Mapped[int | None] = mapped_column(Integer, nullable=True)
     thread_id: Mapped[str | None] = mapped_column(String(MAX_ID_LENGTH), nullable=True)
-    review_comment_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    review_comment_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     path: Mapped[str | None] = mapped_column(String(MAX_ENTITY_KEY_LENGTH), nullable=True)
     line: Mapped[int | None] = mapped_column(Integer, nullable=True)
     side: Mapped[str | None] = mapped_column(String(20), nullable=True)
